refactor: Race 객체 기능 일부를 RaceGame로 분리
Race 객체가 Model의 역할에 충실하도록 일부 기능을 RaceGame 객체로 분리함.
