+++ src/main/java/racingcar/Race.java
@@ -1,10 +1,8 @@
 package racingcar;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 public class Race {
     private final List<Car> cars;
@@ -27,43 +25,11 @@ public class Race {
                 .toList();
     }
 
-    public void playRacing() {
-        for (int i = 0; i < tryCount; i++) {
-            carsMove();
-            OutputView.printRacingResult(cars);
-        }
-
-        OutputView.printWinner(getWinner());
+    public List<Car> getCars() {
+        return cars;
     }
 
-    private void carsMove() {
-        for (Car car : cars) {
-            car.moveCar();
-        }
-    }
-
-    private List<String> getWinner() {
-        int maxPosition = 0;
-        List<String> winner = new ArrayList<>();
-
-        for (Car car : cars) {
-            maxPosition = getMaxPosition(car.getPosition(), maxPosition);
-        }
-
-        for (Car car : cars) {
-            if (car.getPosition() == maxPosition) {
-                winner.add(car.getName());
-            }
-        }
-
-        return winner;
-    }
-
-    private int getMaxPosition(int position, int maxPosition) {
-        if (position >= maxPosition) {
-            return position;
-        }
-
-        return maxPosition;
+    public int getTryCount() {
+        return tryCount;
     }
 }
